requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings==2.0.3  # Added for BaseSettings
openai==1.3.5  # Updated for new assistants API
python-dotenv==1.0.0
torch>=1.7.0
numpy>=1.19.0
pymongo==3.12.3  # Added specific version
motor==2.5.1  # Downgraded to compatible version
slowapi==0.1.8
transformers==4.33.2
sentence-transformers==2.2.2
nest-asyncio==1.5.8  # Added for handling nested event loops
tqdm>=4.46.0
matplotlib>=3.2.1
